Make sure the navbar doesn't jump on resize
The responsive CSS makes the navbar a little taller and swaps out the
menu for a drop-down.  That's nice, but impractical for many Django apps
as it currently stands as they present too much information.

This overrides the default Bootstrap CSS to keep the navbar the same
size during a resize.  This is a feature to revisit in the future, but
will require a little more thought to make sure a useful menu is
presented.